- Implement complete StudentRepo with CRUD operations for student management
- Add validation and error handling to RegistrationRepo methods
- Add UserMessageException for better error reporting with HTTP status codes
- Change registration_no auto-increment from DB to Java-side generation
- Update StudentDetailsModel roll_no from int to long for consistency
- Add separate delete methods for different scenarios (reject vs normal delete)
- Add validation for empty/null fields in update operations
- Add rowsAffected checks for all database operations
- Implement student enrollment status management
- Add helper methods for roll_no and registration_no lookups
- Create mapToStudent method for ResultSet to model mapping
- Add BASE_QUERY for reusable student data retrieval
- Implement filtered read operations (enrolled, not enrolled, by regn_no)basak@gmail.com>

+++ backend/backend/src/main/java/com/studentmanagesystem/backend/model/StudentDetailsModel.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class StudentDetailsModel {
 
-    private int roll_no; // pk, autoincrement
+    private long roll_no; // pk, sql-side autoincrement
 
     private RegistrationModel registration;
     private String subjects;
